Fix build errors: remove duplicate class stubs, escape dollar sign in string, convert Long to Int
- Removed stub PerAppSettings and SnippetManager class definitions from Roadmap.kt (causing 'Redeclaration' errors)
- Escaped dollar sign in Layouts.kt line 80 (string '@#\&-+()/')
- Fixed ClipboardView.kt line 31: convert 0x99FFFFFF to Int for and() function

Build now succeeds with 38 actionable tasks in 14s. APK installed successfully on device.

+++ app/src/main/java/com/openswift/keyboard/ui/ClipboardView.kt
@@ -28,7 +28,7 @@ class ClipboardView @JvmOverloads constructor(
     private val labelPaint = Paint(Paint.ANTI_ALIAS_FLAG).apply {
         textAlign = Paint.Align.LEFT
         textSize = 12f
-        color = theme.keyText.and(0x99FFFFFF)
+        color = theme.keyText.and(0x99FFFFFF.toInt())
     }
     private val deletePaint = Paint(Paint.ANTI_ALIAS_FLAG).apply {
         textAlign = Paint.Align.CENTER
